Fixed time issue and multiple files shenanigans

+++ src/main/java/io/github/abdirashidexe/centroidfinder/FrameData.java
@@ -1,3 +1,3 @@
 package io.github.abdirashidexe.centroidfinder;
 
-public record FrameData(int time, int x, int y) {}
+public record FrameData(double time, int x, int y) {}
